# todoApp/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: Todo App
  version: 0.0.0
...
paths:
...
  /items/{itemId}/attachments:
    get:
      operationId: Attachments_list
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            readOnly: true
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoAttachmentList'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItems.NotFoundErrorResponse'
        4XX:
          description: Something is wrong with you.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Standard4XXResponse'
        5XX:
          description: Something is wrong with me.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Standard5XXResponse'
    post:
      operationId: Attachments_createUrlAttachment_Attachments_createFileAttachment
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            readOnly: true
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItems.NotFoundErrorResponse'
        4XX:
          description: Something is wrong with you.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Standard4XXResponse'
        5XX:
          description: Something is wrong with me.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Standard5XXResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoAttachment'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileAttachmentMultipartRequest'
            encoding:
              contents:
                contentType: '*/*'
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
...
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A machine readable error code
        message:
          type: string
          description: A human readable message
          x-ms-primary-error-message: true
    FileAttachmentMultipartRequest:
      type: object
      properties:
        contents:
          type: string
          format: binary
      required:
        - contents
    Standard4XXResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiError'
      description: Something is wrong with you.
    Standard5XXResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiError'
      description: Something is wrong with me.
...
    TodoAttachment:
      type: object
      required:
        - filename
        - mediaType
        - contents
      properties:
        filename:
          type: string
          maxLength: 255
          description: The file name of the attachment
        mediaType:
          type: string
          description: The media type of the attachment
        contents:
          type: string
          format: byte
          description: The contents of the file
    TodoAttachmentList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodoAttachment'
...
    TodoItems.NotFoundErrorResponse:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - not-found
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: session-id